# apps/web/lib/api/types.ts
import type { components } from "./generated";

export type Summary = components["schemas"]["Summary"];
export type Run = components["schemas"]["Run"];
export type Project = components["schemas"]["Project"];
export type Session = components["schemas"]["Session"];
export type Machine = components["schemas"]["Machine"];